Network.Curl.Hascurl

Description

Simple Haskell bindings to a small subset of libcurl. Most packages available now either seem actively maintained or provide a more feature-rich network request experience, so this package is meant for personal use for when I only need the very basics.

Documentation

get :: String -> IO (Maybe String)

Perform a GET request to the url url. Returns Just cs if the request is successful and the response is cs, and returns Nothing otherwise.